test: add error-summary assertion helpers and normalise template test names
Add assertNoErrors and assertErrorSummary helpers to @hmcts/test-support
(the reference-data-upload template test referenced them but they were
never implemented). Rename the two system-admin template.test.ts files to
the index.njk.test.ts convention and fold the duplicate delete-court-success
locale checks into the rendering test.
